Create EqualVariance_stat
This file contains the scripts used to conduct a levene's test of equality of variances on both multivariate and univariate data for all three communities. It includes generic codes, and codes with the specific workflow we used; filepaths are unique to our directory and should be updated for your directory if you intend to replicate our tests.