Taxonomy keys, OTU tables
This folder has the original OTU table for the 16s and ITS dataset, and a separate file for the taxonomy assignments. In the ITS folder, there's a file that includes the OTU table used for PCORD; sheet 2 in that excel file has the key for matching denovo clusters to the format required for PCORD. In the 16s folder, I included a directory called ArchBac_Dissambiguation which includes excel files in which I filtered the original OTU tables (genus through phylum taxonomic levels) into bacterial and archaeal groups. The key to match taxonomy to the PCORD format is within each of the files (it's a different key for each taxonomic level).