OTU tables, Taxonomy assignments
This has the original OTU tables and Taxonomy assignments (as separate files) for the 16s and ITS libraries. The ITS folder also has a file formatted for PCORD with the required key to match denovo sequences to the format for PCORD. The 16s folder also has a directory containing the excel files used to filter the bacterial from archaeal datasets; there is one file for each of the taxonomic levels (genus - phylum), and each file contains the key for PCORD format OTU to the taxonomic assignment. These files should allow any replication of PCORD analyses, and any indicator species analysis.